Conflicts:
	cdap-docs/reference-manual/source/release-notes.rst

+++ cdap-app-fabric/src/main/java/co/cask/cdap/gateway/handlers/NamespaceHttpHandler.java
@@ -19,19 +19,20 @@ package co.cask.cdap.gateway.handlers;
 import co.cask.cdap.app.store.Store;
 import co.cask.cdap.app.store.StoreFactory;
 import co.cask.cdap.common.conf.Constants;
+import co.cask.cdap.config.PreferencesStore;
 import co.cask.cdap.gateway.auth.Authenticator;
 import co.cask.cdap.gateway.handlers.util.AbstractAppFabricHttpHandler;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.NamespaceMeta;
 import co.cask.http.HttpHandler;
 import co.cask.http.HttpResponder;
+import com.google.common.base.CharMatcher;
 import com.google.gson.JsonSyntaxException;
 import com.google.inject.Inject;
 import org.jboss.netty.handler.codec.http.HttpRequest;
 import org.jboss.netty.handler.codec.http.HttpResponseStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 
 import java.io.IOException;
 import javax.ws.rs.DELETE;
@@ -48,11 +49,14 @@ public class NamespaceHttpHandler extends AbstractAppFabricHttpHandler {
   private static final Logger LOG = LoggerFactory.getLogger(NamespaceHttpHandler.class);
 
   private final Store store;
+  private final PreferencesStore preferencesStore;
 
   @Inject
-  public NamespaceHttpHandler(Authenticator authenticator, StoreFactory storeFactory) {
+  public NamespaceHttpHandler(Authenticator authenticator, StoreFactory storeFactory,
+                              PreferencesStore preferencesStore) {
     super(authenticator);
     this.store = storeFactory.create();
+    this.preferencesStore = preferencesStore;
   }
 
   @GET
@@ -67,24 +71,25 @@ public class NamespaceHttpHandler extends AbstractAppFabricHttpHandler {
   }
 
   @GET
-  @Path("/namespaces/{namespace}")
-  public void getNamespace(HttpRequest request, HttpResponder responder, @PathParam("namespace") String namespace) {
+  @Path("/namespaces/{namespace-id}")
+  public void getNamespace(HttpRequest request, HttpResponder responder,
+                           @PathParam("namespace-id") String namespaceId) {
     try {
-      NamespaceMeta ns = store.getNamespace(Id.Namespace.from(namespace));
+      NamespaceMeta ns = store.getNamespace(Id.Namespace.from(namespaceId));
       if (ns == null) {
-        responder.sendString(HttpResponseStatus.NOT_FOUND, String.format("Namespace %s not found", namespace));
+        responder.sendString(HttpResponseStatus.NOT_FOUND, String.format("Namespace %s not found", namespaceId));
         return;
       }
       responder.sendJson(HttpResponseStatus.OK, ns);
     } catch (Exception e) {
-      LOG.error("Internal error while getting namespace '{}'", namespace, e);
+      LOG.error("Internal error while getting namespace '{}'", namespaceId, e);
       responder.sendStatus(HttpResponseStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
   @PUT
-  @Path("/namespaces")
-  public void create(HttpRequest request, HttpResponder responder) {
+  @Path("/namespaces/{namespace-id}")
+  public void create(HttpRequest request, HttpResponder responder, @PathParam("namespace-id") String namespaceId) {
     NamespaceMeta metadata;
     try {
       metadata = parseBody(request, NamespaceMeta.class);
@@ -97,32 +102,46 @@ public class NamespaceHttpHandler extends AbstractAppFabricHttpHandler {
       return;
     }
 
-    String name = metadata.getName();
-    // name cannot be null or empty.
-    if (name == null || name.isEmpty()) {
-      responder.sendString(HttpResponseStatus.BAD_REQUEST, "Namespace name cannot be null or empty.");
+    if (!isValid(namespaceId)) {
+      responder.sendString(HttpResponseStatus.BAD_REQUEST,
+                           "Namespace id can contain only alphanumeric characters, '-' or '_'.");
       return;
     }
+
+    if (isReserved(namespaceId)) {
+      responder.sendString(HttpResponseStatus.BAD_REQUEST,
+                           String.format("'%s' and '%s' are reserved namespace ids.",
+                                         Constants.DEFAULT_NAMESPACE,
+                                         Constants.SYSTEM_NAMESPACE));
+      return;
+    }
+
     // displayName and description could be null
     String displayName = metadata.getDisplayName();
     if (displayName == null || displayName.isEmpty()) {
-      displayName = name;
+      displayName = namespaceId;
     }
     String description = metadata.getDescription();
     if (description == null) {
       description = "";
     }
 
-    try {
-      NamespaceMeta existing = store.createNamespace(new NamespaceMeta.Builder().setName(name)
-                                                       .setDisplayName(displayName).setDescription(description)
-                                                       .build());
-      if (existing == null) {
-        responder.sendStatus(HttpResponseStatus.OK);
-      } else {
-        responder.sendString(HttpResponseStatus.CONFLICT, String.format("Namespace %s already exists.", name));
-      }
+    NamespaceMeta.Builder builder = new NamespaceMeta.Builder();
+    builder.setId(namespaceId)
+      .setDisplayName(displayName)
+      .setDescription(description)
+      .build();
 
+    try {
+      NamespaceMeta existing = store.createNamespace(builder.build());
+      // make the API idempotent, but send appropriate response
+      String response;
+      if (existing == null) {
+        response = String.format("Namespace '%s' created successfully.", namespaceId);
+      } else {
+        response = String.format("Namespace '%s' already exists.", namespaceId);
+      }
+      responder.sendString(HttpResponseStatus.OK, response);
     } catch (Exception e) {
       LOG.error("Internal error while creating namespace.", e);
       responder.sendStatus(HttpResponseStatus.INTERNAL_SERVER_ERROR);
@@ -130,10 +149,20 @@ public class NamespaceHttpHandler extends AbstractAppFabricHttpHandler {
   }
 
   @DELETE
-  @Path("/namespaces/{namespace}")
-  public void delete(HttpRequest request, HttpResponder responder, @PathParam("namespace") String namespace) {
+  @Path("/namespaces/{namespace-id}")
+  public void delete(HttpRequest request, HttpResponder responder, @PathParam("namespace-id") String namespace) {
+    if (isReserved(namespace)) {
+      responder.sendString(HttpResponseStatus.FORBIDDEN,
+                           String.format("'%s', '%s' are reserved namespace ids",
+                                         Constants.DEFAULT_NAMESPACE,
+                                         Constants.SYSTEM_NAMESPACE));
+      return;
+    }
     Id.Namespace namespaceId = Id.Namespace.from(namespace);
     try {
+      // Delete Preferences associated with this namespace
+      preferencesStore.deleteProperties(namespace);
+
       // Store#deleteNamespace already checks for existence
       NamespaceMeta deletedNamespace = store.deleteNamespace(namespaceId);
       if (deletedNamespace == null) {
@@ -145,5 +174,18 @@ public class NamespaceHttpHandler extends AbstractAppFabricHttpHandler {
       LOG.error("Internal error while deleting namespace.", e);
       responder.sendStatus(HttpResponseStatus.INTERNAL_SERVER_ERROR);
     }
+  }
+
+  private boolean isValid(String namespaceId) {
+    // TODO: This is copied from StreamVerification in app-fabric as this handler is in data-fabric module.
+    return CharMatcher.inRange('A', 'Z')
+      .or(CharMatcher.inRange('a', 'z'))
+      .or(CharMatcher.is('-'))
+      .or(CharMatcher.is('_'))
+      .or(CharMatcher.inRange('0', '9')).matchesAllOf(namespaceId);
+  }
+
+  private boolean isReserved(String namespaceId) {
+    return Constants.DEFAULT_NAMESPACE.equals(namespaceId) || Constants.SYSTEM_NAMESPACE.equals(namespaceId);
   }
 }
